stock-analysis/requirements.txt
streamlit
pandas
yfinance
plotly
ta
pandas_ta
python-dateutil
